requirements.txt
flask
pyyaml
gunicorn
python-dotenv
#pip install -r requirements.txt
